【开发自提单】【BeiMing-K】【大数据】【BoostIO】【安全专项】uint64强转uint16,高位截断


Match-id-d61261dec71dfff00c8abf7ca73f464a1c43e13e

+++ src/cache/read/rcache.cpp
@@ -17,7 +17,7 @@ using namespace ock::bio;
 
 static constexpr uint32_t RCACHE_FLOW_PREFIX_START = RCACHE_FLOW_MEM_META_PREFIX;
 
-RCache::RCache(uint64_t ptId, uint64_t ptv, uint16_t diskId, uint32_t workIndex)
+RCache::RCache(uint16_t ptId, uint64_t ptv, uint16_t diskId, uint32_t workIndex)
     : mFlowId(0), mPtId(ptId), mPtv(ptv), mDiskId(diskId), mWorkIndex(workIndex)
 {
     for (int32_t tier = 0; tier < READ_CACHE_TIER_BUTT; tier++) {
